feat: 침묵 시간 & 목소리 크기 프로세스 분리 (#63)
