+++ src/main/java/com/example/Namanba/evaluation/entity/Category.java
@@ -10,7 +10,9 @@ public enum Category {
     GAZE("gaze"),
     EXPRESSION("expression"),
 
-    SILENCEDURATION("silenceDuration");
+    SILENCEDURATION("silenceDuration"),
+
+    VOICEVOLUME("voiceVolume");
 
     private final String value;
 }
