+++ src/main/java/com/example/Namanba/evaluation/service/EvaluationDomainService.java
@@ -68,6 +68,13 @@ public class EvaluationDomainService {
     }
 
     @Transactional
+    public void evaluateVoiceVolume(Interview interview, AudioEvaluationDto audioEvaluationDto){
+        Evaluation evaluation = evaluationAdaptor.findByInterview(interview);
+        evaluation.assignVoiceVolume(audioEvaluationDto.getVoiceVolume(), audioEvaluationDto.getVoiceVolumeMessage());
+        evaluationAdaptor.save(evaluation);
+    }
+
+    @Transactional
     public void evaluateExpression(Interview interview, ExpressionEvaluationDto expressionEvaluationDto){
         Evaluation evaluation = evaluationAdaptor.findByInterview(interview);
         // expression 값을 검증하고 기본값 설정
